+++ src/api/api.js
@@ -102,20 +102,12 @@ class MinkApi {
     return response.data;
   }
 
-  /**
-   * @returns {job_status, message, status} job_status can be: none, syncing_corpus,
-   *   waiting, annotating, done_annotating, syncing_results, done, error, aborted.
-   */
-  async checkStatus(corpusId) {
-    const response = await this.axios.get("check-status", {
+  /** @see https://ws.spraakbanken.gu.se/ws/mink/api-doc#tag/Process-Corpus/operation/resourceinfo */
+  async resourceInfo(corpusId) {
+    const response = await this.axios.get("resource-info", {
       params: { corpus_id: corpusId },
     });
     return response.data;
-  }
-
-  async checkStatusAll() {
-    const response = await this.axios.get("check-status");
-    return response.data.jobs;
   }
 
   async runSparv(corpusId) {